% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastpnorm}
\alias{fastpnorm}
\alias{fastpnorm_preallocated}
\title{Piecewise Linear Interpolation for pnorm}
\usage{
fastpnorm(q, precise_tail = FALSE, use_cubic = FALSE, n_threads = 1L)

fastpnorm_preallocated(
  q,
  p,
  precise_tail = FALSE,
  use_cubic = FALSE,
  n_threads = 1L
)
}
\arguments{
\item{q}{vector of quantiles.}

\item{precise_tail}{logical for whether to use \code{\link{pnorm}} in
the tail.}

\item{use_cubic}{true if monotone cubic interpolation using Fritsch–Carlson
method should be used.}

\item{n_threads}{number of threads to use.}

\item{p}{pector vector outputs (probabilities). Needs to have the same
length as \code{q}.}
}
\description{
Performs linear interpolation of \code{\link{pnorm}}. See the README
at \url{https://github.com/yixuan/fastncdf} for details.

The \code{fastpnorm_preallocated} version is faster if one has
vector of the same length already.
}
\examples{
x <- seq(-6, 6, by = 1e-6)
system.time(y <- pnorm(x))
system.time(fasty <- fastpnorm(x))
all.equal(y, fasty) # tiny error
}
