% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastpnorm}
\alias{fastpnorm}
\alias{fastpnorm_preallocated}
\title{Piecewise Linear Interpolation for pnorm}
\usage{
fastpnorm(q)

fastpnorm_preallocated(p, q)
}
\arguments{
\item{q}{vector of quantiles.}

\item{p}{pector vector outputs (probabilities). Needs to have the same
length as \code{q}.}
}
\description{
Performs linear interpolation of \code{\link{pnorm}}. See the README
at \url{https://github.com/yixuan/fastncdf} for details.

The \code{fastpnorm_preallocated} version is faster if one has
vector of the same length already.
}
\examples{
x <- seq(-6, 6, by = 1e-6)
system.time(y <- pnorm(x))
system.time(fasty <- fastpnorm(x))
all.equal(y, fasty) # tiny error
}
